Implement JWT-based session mechanism and SMTP authentication provider
Added a new authentication provider allowing sign in with SMTP, and updated the session handling to use JSON Web Tokens (JWT) instead of simple random strings. Also introduced minor changes to the Discord provider to accommodate the use of JWT. Libraries necessary for validating and generating JWT have been added to the list of project dependencies. Sign In page was updated to include form inputs for users to enter their email and password, and sign in using the new SMTP provider.